requirements.txt
PyQt5
paddlepaddle
paddleocr
opencv-python
Pillow
numpy
PyYAML
Shapely
imgaug